util: Use uint64 frequencies in bands

+++ pkg/band/ru_864_870.go
@@ -143,7 +143,7 @@ func init() {
 		Beacon: Beacon{
 			DataRateIndex:    3,
 			CodingRate:       "4/5",
-			BroadcastChannel: func(_ float64) uint32 { return 869100000 },
+			BroadcastChannel: func(_ float64) uint64 { return 869100000 },
 			PingSlotChannel:  uint64Ptr(868900000),
 		},
